Add per-kernel data data statements to compute_plank_source
This required a small bugfix; the first argument to interpolate3D was
being passed as in inline array constructor. PGI then tried to generate
an implicit copyin for an array temporary to handle this. In order to
prevent the implicit copyin (and the array temporary), the array was
moved to a variable scaling_array = (/1._wp,1._wp/), and this array
was given an explicit copyin statement.