Fix data statements for OpenACC
Fix some data statements that were causing runtime problems with OpenACC with PGI.
This mostly just removes arrays from the explicit data movement that did not need
to persist between kernels (only used in one kernel), so the compiler instead
inserts the data movement directives automatically. For some reason this fixed
the runtime issues (still not sure why), and should not change the data movement
at runtime or the answers (flux_compute and gas_optics tests still pass). HOWEVER,
the tests at this commit will NOT compile with openACC flags with PGI, due to a link
error when linking against mo_gas_optics_kernels; the compiler complains about:

nvlink error   : Undefined reference to '__pgi_ldg_unknown' in 'librrtmgp.a:mo_gas_optics_kernels.o'

It turns out that replacing all occurrences of `logical(wl)` with `logical` (removing
C interoperability) makes this go away, but we did not want to do this at this point
because it would BREAK C compatibility even for non-OpenACC builds. Work is
on-going to identify the source of this and a possible workaround that does not
breka C compatibility.