Fix compiler flags for PGI/OpenACC bug
Turn off optimization for PGI/OpenACC builds. There is a compiler bug
that seems to be related to using C-compatible logical arrays in
subroutine arguments with optimization. Turning off optimization
allows the code to compile. This should be considered a *temporary*
workaround until the PGI/NVIDIA folks can provide a bug fix (ticket
submitted for this).