Use CUDA managed memory for OLCF PGI builds
Use CUDA managed memory for GPU builds on OLCF Summit. This should not
be needed for the kernels themselves, but becomes necessary when we try to
put data on the GPU at a higher level. There are still some bugs to work
out if we want to explicitly handle the data from end to end, but using
managed memory works around these with minimal intrusion to the user for
now.