add single optimal angle coefficients to LW coefficient file
This full set of code will not be ported to master but is a branch commit until I can build the LW coefficients file within the "build" infrastructure.

New name of coefficients file has an "a" appended.

Note: coefficients file in flux_compute.ini points to trial version of amended coefficients file.