Add code for computing optimal single angle for no-scattering calculations (#66)
Adding code and validation tests for using an optimal angle for longwave no-scattering calculations. Routine gas_optics%compute_optimal_angles() returns the secant of the propagation angle for each g-point that minimizes errors with respect to high angular-resolution calculations. 

We are still waiting for data so, for now, the routine returns the same angle as is used by default. 

OpenACC implementation includes help from Dmitry Alexeev of Nvidia. 

Co-authored-by: Robert Pincus <Robert.Pincus@colorado.edu>