Further clean up of test build
Move scripts out of build directory, and use build as a temporary
directory within which to build and run the tests. Thus, build will
contain all the temporary files for a test that are not version
controlled, including all of the netcdf files needed to run a test, and
all of the object files.