Add computation for net band-by-band fluxes
Add missing computation for net band-by-band fluxes. A previous PR added
the FluxesByband class to keep track of spectral fluxes, but we
neglected to properly set the resulting net fluxes (only the up and
down). This also adds a test to the allsky test code to make sure that
the sum of the computed byband fluxes matches the sum of the broadband
fluxes as a sanity check.

[BFB]