Feature streamline LW (#101)
Longwave code is streamlined 1) to compute spectrally-integrated Jacobians with respect to surface temperature only when this output is requested, and 2) to combine the no-scattering and rescaled-for-scattering approaches as much as possible. Argument lists to some kernels have changed but arguments to the front-end have not. nvfortran is failing and needs debugging. OpenACC implementation may not be optimal.