Make to build, run tests,  check results (#106)
Makefiles are added and amended to build libraries, run tests, and verify results using "make libs", "make tests", "make check" respectively in the top-level directory. Environmental variables set the Fortran compiler and flags and the locations of the netCDF libraries and include files.

The CSCS project used for CI is also updated.