Feature: reordered gas optics (#148)
* Reorder data, avoid transposition in RRTMGP (#145)

This refactoring changes the order of the RRTMGP lookup tables. Optical depth and source functions are now computed for all columns with the g-point dimensions last. This may be harder to vectorize but avoids the transposition of very large final matrices. The GPU implementation also fuses loops in the computation of Planck source function for efficiency. 

OpenMP implementation is no broken - we test with Cray CCE compiler with fails with an internal error. 

Co-authored-by: Erik Pfister <pfister@dkrz.de>
Co-authored-by: Dmitry Alexeev <dalexeev@nvidia.com>
Co-authored-by: Panos Adamidis <k202061@mlogin100.hpc.dkrz.de>
Co-authored-by: Robert Pincus <Robert.Pincus@columbia.edu>