Extend Levante CI with NAG (#260)
This adds CI tests with the NAG compiler on Levante with compiler version and flags currently used to build ICON by default.
 
Both DP and SP floating models are tested, as are default and accelerator kernels, but not all possible combinations. Accelerator kernels fail with a run-time error and are marked as experimental.